+++ crates/rpc/src/method/trace_transaction.rs
@@ -124,7 +124,7 @@ pub async fn trace_transaction(
                 context.chain_id,
                 header,
                 None,
-                context.config.custom_versioned_constants,
+                context.config.versioned_constants_map,
                 context.contract_addresses.eth_l2_token_address,
                 context.contract_addresses.strk_l2_token_address,
             );